+++ package.json
@@ -13,6 +13,7 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
+    "install": "^0.13.0",
     "jspdf": "^2.5.1",
     "pinia": "^2.1.7",
     "vue": "^3.4.15",
