+++ src/router/index.ts
@@ -7,12 +7,9 @@ import ContactView from "../views/ContactView.vue"
 import ActivitiesView from "../views/ActivitiesView.vue"
 import ComprarUnoView from "../views/ComprarUno.vue"
 import ComprarDosView from "../views/ComprarDos.vue"
-import HomeAdminView from "../views/HomeAdminView.vue"
+
 import AdminPanelView from "../views/AdminPanelView.vue"
 import LoginView from "../views/LoginView.vue"
-import LoginAdminView from "../views/LoginAdminView.vue"
-import LoginUserView from "../views/LoginUserView.vue"
-
 
 const router = createRouter({
     history: createWebHistory(import.meta.env.BASE_URL),
@@ -21,27 +18,13 @@ const router = createRouter({
             path: '/',
             name: 'Login',
             component: LoginView
-        },
-        {
-            path: '/LoginUser',
-            name: 'LoginUser',
-            component: LoginUserView
-        },
-        {
-            path: '/LoginAdmin',
-            name: 'LoginAdmin',
-            component: LoginAdminView
-        },     
+        },    
         {
             path: '/Home',
             name: 'Home',
             component: HomeView
         },
-        {
-            path: '/HomeAdmin',
-            name: 'HomeAdmin',
-            component: HomeAdminView
-        },
+       
         {
             path: '/AdminPanel',
             name: 'AdminPanel',
