+++ src/views/HomeView.vue
@@ -18,33 +18,14 @@
         </div>
       </section>
       <article>
-        <section class="poster-container">
-          <div v-for="obra in store.obras" :key="obra.obraID" class="show-poster">
-            <div class="show-poster__image">
-              <img :src="obra.imagenes.split(',')[0]" alt="Imagen de la obra" />
-            </div>
-            <div class="show-poster__details">
-              <h3 class="show-poster__details__title">{{ obra.nombre }}</h3>
-              <RouterLink :to="{ path: '/Function/' + obra.obraID }" class="show-poster__button">
-                {{ $t('HomeAdmin.text4') }}
-              </RouterLink>
-            </div>
-          </div>
-        </section>
+        <CardComponent></CardComponent>
       </article>
     </main>
   </body>
 </template>
 
 <script setup lang="ts">
-import { useListadoObrasStore } from '../store/Obra-Store'
-import { ref, onMounted } from 'vue'
-import { RouterLink } from 'vue-router'
-const store = useListadoObrasStore()
-
-onMounted(() => {
-  store.cargarObras()
-})
+import CardComponent from '../components/CardComponent.vue'
 </script>
 
 <style scoped>
